+++ rascil/processing_components/skycomponent/operations.py
@@ -309,21 +309,20 @@ def find_skycomponents(
         im["pixels"].data.shape[0]
     )
     segments = segmentation.detect_sources(
-        image_sum, threshold, npixels=npixels, filter_kernel=kernel
+        image_sum, threshold, npixels=npixels, kernel=kernel
     )
     if segments == None:
         raise ValueError("find_skycomponents: Failed to find any components")
 
     log.info("find_skycomponents: Identified %d segments" % segments.nlabels)
 
-    # Now compute source properties for all polarisations and frequencies
-    comp_tbl = [
+    comp_catalog = [
         [
-            segmentation.source_properties(
+            segmentation.SourceCatalog(
                 im["pixels"].data[chan, pol],
                 segments,
-                filter_kernel=kernel,
-            ).to_table()
+                kernel=kernel,
+            )
             for pol in [0]
         ]
         for chan in range(im.image_acc.nchan)
@@ -331,7 +330,7 @@ def find_skycomponents(
 
     def comp_prop(comp, prop_name):
         return [
-            [comp_tbl[chan][pol][comp][prop_name] for pol in [0]]
+            [comp_catalog[chan][pol][comp].__getattribute__(prop_name) for pol in [0]]
             for chan in range(im.image_acc.nchan)
         ]
 
@@ -341,8 +340,8 @@ def find_skycomponents(
         # Get flux and position. Astropy's quantities make this
         # unnecessarily complicated.
         flux = numpy.array(comp_prop(segment, "max_value"))
-        xs = u.Quantity(list(map(u.Quantity, comp_prop(segment, "maxval_xpos"))))
-        ys = u.Quantity(list(map(u.Quantity, comp_prop(segment, "maxval_ypos"))))
+        xs = u.Quantity(list(map(u.Quantity, comp_prop(segment, "maxval_xindex"))))
+        ys = u.Quantity(list(map(u.Quantity, comp_prop(segment, "maxval_yindex"))))
 
         sc = pixel_to_skycoord(xs, ys, im.image_acc.wcs, 0)
         ras = sc.ra