+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 master
 ------
 
+* Update rascil.processing_components.skycomponent.operations.find_skycomponents to use photutils==1.4.0 ([MR322](https://gitlab.com/ska-telescope/external/rascil/-/merge_requests/322))
 * Bug fix for exporting measurement set file with multi-block visibility ([MR315](https://gitlab.com/ska-telescope/external/rascil/-/merge_requests/315)
 
 0.6.0